silabs-multiprotocol: Add amd64 support (#2757)
* Support AMD64 architecture through QEMU user emulation for zigbeed

Add support for AMD64 (x86-64) by using QEMU. All components other than
zigbeed are running natively using x86-64. However, since zigbeed comes
only with armv7 and aarch64 support, we have to use QEMU emulation to
use it on x86-64.

* Use non-static variant

It is much smaller and seems to work well for this use case

* Drop stdbuf operation

This causes problems with QEMU emulation. It does not seem to be
required for zigbeed.

* Bump version

* Add missing patch file

* Convert stdout print to comment

* Bump Multicast table size to 16

* Extend changelog

Co-authored-by: Pascal Vizeli <pvizeli@syshack.ch>